BogoBots/requirements.txt
mysql-connector-python==8.3.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
streamlit==1.39.0
streamlit-extras==0.4.0
xlrd==2.0.1
plotly==5.22.0
openai==1.30.4
langchain==0.2.1
langchain_core==0.2.3
langchain_openai==0.1.8
langchain_community==0.2.1
langchain_huggingface==0.0.1
langchain_milvus==0.1.0
langgraph==0.0.55
pymilvus==2.4.3
mysqlclient==2.2.4
ebooklib==0.17.1
httpx==0.27.2
setuptools==81.0.0
# AI Hub dependencies
feedparser==6.0.11
python-dateutil==2.9.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.2.1
tweepy==4.14.0